+++ liteidex/macosx-deploy.sh
@@ -7,5 +7,6 @@ cp -v LGPL_EXCEPTION.TXT liteide/bin
 cp -r -v deploy/* liteide/bin/LiteIDE.app/Contents/Resources
 cp -r -v os_deploy/macosx/* liteide/bin/LiteIDE.app/Contents/Resources
 
-cp -v bin/* liteide/bin/LiteIDE.app/Contents/MacOS
+cp -v bin/liteide_stub liteide/bin/LiteIDE.app/Contents/MacOS
+cp -v bin/gocode liteide/bin/LiteIDE.app/Contents/MacOS
 cp -v src/liteapp/Info.plist liteide/bin/LiteIDE.app/Contents